refactor: FIleInputコンポーネントを汎用的にする
