+++ frontend/src/features/course/hooks/useCourse.ts
@@ -97,7 +97,7 @@ export const useCreateCourseWithRelation = () => {
     useGetApi<CourseWithSections[]>('/courses')
 
   const createWithRelationCourse = useCallback(
-    async (params: CourseWithSections) => {
+    async (params: object) => {
       try {
         const newCourse = await postApi<CourseWithSections>(
           '/courses/bulk',
@@ -111,10 +111,8 @@ export const useCreateCourseWithRelation = () => {
       } catch (e) {
         console.error(e)
         if (e instanceof Error) {
-          throw `コースの作成に失敗しました。\n以下が原因である可能性があるので、内容を確認し再度お試しください。\n
+          throw `コースの作成に失敗しました。\n以下のような原因が考えられます。内容を確認し再度お試しください。\n
 - コースの名前(name)・制作者(author)・所属(organization)の組み合わせが既に存在する。
-- 必須項目が抜けている。
-- 項目の値に誤りがある。
 \n 以下はエラー内容です${e.message}`
         }
         throw JSON.stringify(e)
