+++ frontend/src/features/course/index.ts
@@ -1,5 +1,4 @@
 export * from './components/CreateCourseButton'
 export * from './components/UpdateCourseButton'
 export * from './components/CourseTable'
-export * from './components/CourseFileInput'
 export * from './hooks/useCourse'
